chore: enable local server startup
