+++ memory.js
@@ -1,0 +1,9 @@
+let eventsMemory = [];
+export const getEventsMemory = () => eventsMemory;
+export const setEventsMemory = (events) => {
+  eventsMemory = events;
+};
+
+export const notificationsMemory = [];
+export const mockVolunteers = [];
+export const staticEvents = [];
