+++ utils/notifications.js
@@ -1,0 +1,15 @@
+import { query, USE_DB } from "../config/db.js";
+import { notificationsMemory } from "../memory.js";
+
+export const addNotification = async (userId, message) => {
+  if (!USE_DB) {
+    const n = { id: Date.now(), userId: Number(userId), message, read: false };
+    notificationsMemory.push(n);
+    return n;
+  }
+  await query(
+    "INSERT INTO notifications (userId, message, is_read) VALUES (?, ?, ?)",
+    [Number(userId), message, 0]
+  );
+  return { id: undefined, userId: Number(userId), message, read: false };
+};
